Add example component for smoke testing release candidates
