+++ example/package.json
@@ -33,6 +33,7 @@
     "axios": "^0.19.0",
     "emotion-theming": "^10.0.19",
     "react": "^16.9.0",
-    "react-dom": "^16.9.0"
+    "react-dom": "^16.9.0",
+    "use-axios-client": "latest"
   }
 }
