+++ example/src/compiledUsage/index.tsx
@@ -1,34 +1,2 @@
-import React from 'react';
-import Container from '../Container';
-import Heading from '../Heading';
-import TextBlock from '../TextBlock';
-import { useAxios } from '../../../bin';
-
-interface Data {
-  data: {
-    name: string;
-    color: string;
-  };
-}
-
-export default () => {
-  const { data, error, loading } = useAxios<Data>('https://reqres.in/api/things/1');
-
-  return (
-    <Container>
-      <Heading>Compiled Usage</Heading>
-
-      <TextBlock>
-        {loading && 'Loading...'}
-        {error && error}
-        {data && !loading && (
-          <div>
-            {data.data.name}
-            {': '}
-            {data.data.color}
-          </div>
-        )}
-      </TextBlock>
-    </Container>
-  );
-};
+export { default as CompiledLocal } from './CompiledLocal';
+export { default as CompiledNpm } from './CompiledNpm';
