+++ example/src/pages/CompiledUsagePage.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { Flex } from '@chakra-ui/core';
 import { RouteComponentProps } from '@reach/router';
-import CompiledUsage from '../compiledUsage';
+import { CompiledLocal, CompiledNpm } from '../compiledUsage';
 
 const CompiledUsagePage: React.FC<RouteComponentProps> = () => (
   <Flex wrap="wrap">
-    <CompiledUsage />
+    <CompiledLocal />
+    <CompiledNpm />
   </Flex>
 );
